Add files via upload
A2_task is a main module without data & modifier lock. But, it has min latency
A2_task_release is like a previous, but it has +1 clk latency for data input & modifier. Potentially it is more reliable.
A2_task_tb is a testing unit with vary and max load, work with both modules.
A2_task_top is a wrapper for Quartus synth
make_sim is a script for Modelsim